chore: move interfaces to global interface file
